+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Installer for `yazi` and my `yazi` cfg files
     - Wired in the `yazi` installer into install and check process
+    - Updated to new repo with debian/ubuntu specific versions
+    - Also check for static binaries in `/usr/local/bin` and removes them
 - New depdencies to fix build errors in `Trixie`
 - Debian version package awareness to ``00-depdencies.sh`
     - Stops failed install errors as I tried to install all versions of some pkgs