+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Jun 2026
 
-- Fixed: `install.sh` script
-    - Was failing on `trixie` due to issues handling `triie-backports` repo
+- Fixed:
+    - `install.sh` script
+        - Was failing on `trixie` due to issues handling `triie-backports` repo
+    - `install.sh` wasn't showing CLI options with `-h/--help`
 
 ## May 2026