+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - `install.sh` script
         - Was failing on `trixie` due to issues handling `triie-backports` repo
     - `install.sh` wasn't showing CLI options with `-h/--help`
+    - Verified `update.sh` follows this fix and added more defensive code
 
 ## May 2026